feat(demo): replace command-bar with aurora theme
Drops the command-palette mock and coral accent in favour of a more
generic dark developer-doc deck — black canvas, single violet glow,
rounded card shells, monospace metadata. Demo content swaps the kbd
shortcuts for a stat-card grid (cold start / memory / breaking changes).
